+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.0
+* Added Character Sheet mode (Change sheet of characters!)
+* Added Permanent Private Whiteboard (Button in Journal Notes controls)
+
 ## 1.0.2
 * Fixed languages support (Thanks @sladecraven)